refactor(pdf-service): move createPdfKitDocument out of domains folder
