+++ app/domains/fluggastrechte/services/pdf/sections/__test__/createBankInformation.test.ts
@@ -7,7 +7,7 @@ import {
   FONTS_BUNDESSANS_REGULAR,
   PDF_HEIGHT_SEIZE,
   PDF_MARGIN_HORIZONTAL,
-} from "../../createPdfKitDocument";
+} from "~/services/pdf/createPdfKitDocument";
 import { createBankInformation } from "../createBankInformation";
 
 describe("createBankInformation", () => {
