+++ app/domains/fluggastrechte/services/pdf/sections/createPageNumber.ts
@@ -3,7 +3,7 @@ import {
   FONTS_BUNDESSANS_REGULAR,
   PDF_HEIGHT_SEIZE,
   PDF_MARGIN_HORIZONTAL,
-} from "../createPdfKitDocument";
+} from "~/services/pdf/createPdfKitDocument";
 
 export const createPageNumber = (
   doc: typeof PDFDocument,
