+++ app/domains/fluggastrechte/services/pdf/sections/firstPage/__test__/createStatementClaim.test.ts
@@ -5,7 +5,7 @@ import {
 } from "tests/factories/mockPdfKit";
 import type { FluggastrechtContext } from "~/domains/fluggastrechte/formular/context";
 import { getCompensationPayment } from "~/services/airports/getCompensationPayment";
-import { PDF_MARGIN_HORIZONTAL } from "../../../createPdfKitDocument";
+import { PDF_MARGIN_HORIZONTAL } from "~/services/pdf/createPdfKitDocument";
 import {
   createStatementClaim,
   STATEMENT_CLAIM_AGREEMENT_SENTENCE,
