+++ app/domains/fluggastrechte/services/pdf/sections/firstPage/claimData/addAirlineDetails.ts
@@ -1,11 +1,11 @@
 import type PDFDocument from "pdfkit";
 import type { FluggastrechtContext } from "~/domains/fluggastrechte/formular/context";
 import { getAirlineNameByIataCode } from "~/services/airlines/getAirlineNameByIataCode";
-import { SEPARATOR } from "./addPlaintiffDetails";
 import {
   FONTS_BUNDESSANS_BOLD,
   FONTS_BUNDESSANS_REGULAR,
-} from "../../../createPdfKitDocument";
+} from "~/services/pdf/createPdfKitDocument";
+import { SEPARATOR } from "./addPlaintiffDetails";
 
 export const addAirlineDetails = (
   doc: typeof PDFDocument,
