+++ app/domains/fluggastrechte/services/pdf/sections/firstPage/createLocalCourtAndDate.ts
@@ -1,11 +1,11 @@
 import type PDFDocument from "pdfkit";
 import type { FluggastrechtContext } from "~/domains/fluggastrechte/formular/context";
 import { getCourtByStartAndEndAirport } from "~/domains/fluggastrechte/services/getCourtByStartAndEndAirport";
-import { today, toGermanDateFormat } from "~/util/date";
 import {
   FONTS_BUNDESSANS_BOLD,
   FONTS_BUNDESSANS_REGULAR,
-} from "../../createPdfKitDocument";
+} from "~/services/pdf/createPdfKitDocument";
+import { today, toGermanDateFormat } from "~/util/date";
 
 export const CREATION_PDF_TEXT = "Erstellt am:";
 export const TO_THE_COURT_TEXT = "An das";
