+++ app/domains/fluggastrechte/services/pdf/sections/reason/factsOfCases/addFlightDetails.ts
@@ -4,7 +4,7 @@ import { getAirportNameByIataCode } from "~/services/airports/getAirportNameByIa
 import {
   FONTS_BUNDESSANS_BOLD,
   FONTS_BUNDESSANS_REGULAR,
-} from "../../../createPdfKitDocument";
+} from "~/services/pdf/createPdfKitDocument";
 
 export const BOOKING_NUMBER_TEXT = "Buchungsnummer: ";
 export const FLIGHT_NUMBER_TEXT = "Flugnummer: ";
