+++ app/domains/fluggastrechte/services/pdf/sections/reason/factsOfCases/createFactsOfCases.ts
@@ -1,12 +1,12 @@
 import type PDFDocument from "pdfkit";
 import type { FluggastrechtContext } from "~/domains/fluggastrechte/formular/context";
+import { FONTS_BUNDESSANS_BOLD } from "~/services/pdf/createPdfKitDocument";
 import { addCompensationAmount } from "./addCompensationAmount";
 import { addDetailedReason } from "./addDetailedReason";
 import { addFlightDetails } from "./addFlightDetails";
 import { addReason } from "./addReason";
 import { addTable } from "./table/addTable";
 import { COLUMN_HEIGHT } from "./table/tableConfigurations";
-import { FONTS_BUNDESSANS_BOLD } from "../../../createPdfKitDocument";
 
 export const FACTS_OF_CASES_TEXT = "I. Sachverhalt";
 const MARGIN_TOP = 10;
